backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

torch
torchvision
pillow
scikit-learn
joblib
numpy
